Implementação de método para salvar amostra e listar na TableView Diversos outros ajustes de layout

+++ src/main/java/com/servicos/estatica/resicolor/lab/controller/Screen1Controller.java
@@ -70,7 +70,7 @@ public class Screen1Controller implements Initializable, ControlledScreen {
 		chartTemp.setData(plotValuesList);
 	}
 
-	public void plotTemp(Double temp) {
+	public void plotTemp(Integer temp) {
 		final XYChart.Data<String, Number> data = new XYChart.Data<>(dataHoraFormatter.format(LocalDateTime.now()),
 				temp);
 		Node mark = new HoverDataChart(0, temp);
